move parameter values logic to own utils file
