+++ frontend/src/metabase/meta/Card.js
@@ -1,10 +1,10 @@
-import {
-  getTemplateTagParameters,
-  normalizeParameterValue,
-  getValuePopulatedParameters,
-} from "metabase/meta/Parameter";
+import { getTemplateTagParameters } from "metabase/meta/Parameter";
 import { parameterToMBQLFilter } from "metabase/parameters/utils/mbql";
 import { getParameterTargetField } from "metabase/parameters/utils/targets";
+import {
+  normalizeParameterValue,
+  getValuePopulatedParameters,
+} from "metabase/parameters/utils/parameter-values";
 
 import * as Query from "metabase/lib/query/query";
 import * as Q_DEPRECATED from "metabase/lib/query"; // legacy
