+++ frontend/src/metabase/query_builder/selectors.js
@@ -14,7 +14,7 @@ import {
 } from "metabase/visualizations";
 import { getComputedSettingsForSeries } from "metabase/visualizations/lib/settings/visualization";
 import { getValueAndFieldIdPopulatedParametersFromCard } from "metabase/meta/Card";
-import { normalizeParameterValue } from "metabase/meta/Parameter";
+import { normalizeParameterValue } from "metabase/parameters/utils/parameter-values";
 
 import Utils from "metabase/lib/utils";
